refactor:udpate queue protocol api * move GetCryptoList/GetAddressDrivers to rpc client
