+++ client/queueprotocol.go
@@ -10,11 +10,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/33cn/chain33/common/address"
-
 	"github.com/33cn/chain33/common"
-
-	"github.com/33cn/chain33/common/crypto"
 
 	"github.com/33cn/chain33/common/log/log15"
 
@@ -1050,26 +1046,6 @@ func (q *QueueProtocol) GetConfig() *types.Chain33Config {
 		panic("Chain33Config is nil")
 	}
 	return cfg
-}
-
-// GetCryptoList 获取加密算法列表
-func (q *QueueProtocol) GetCryptoList() *types.CryptoList {
-	names, ids := crypto.GetCryptoList()
-	list := &types.CryptoList{Cryptos: make([]*types.Crypto, len(names))}
-	for i, name := range names {
-		list.Cryptos[i] = &types.Crypto{Name: name, TypeID: ids[i]}
-	}
-	return list
-}
-
-// GetAddressDrivers 获取已注册地址插件
-func (q *QueueProtocol) GetAddressDrivers() *types.AddressDrivers {
-	drivers := address.GetDriverList()
-	list := &types.AddressDrivers{Drivers: make([]*types.AddressDriver, 0, len(drivers))}
-	for id, driver := range drivers {
-		list.Drivers = append(list.Drivers, &types.AddressDriver{Name: driver.GetName(), TypeID: id})
-	}
-	return list
 }
 
 // SendDelayTx send delay transaction to mempool
