+++ client/queueprotocol_test.go
@@ -9,10 +9,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/33cn/chain33/common/address"
-
-	"github.com/33cn/chain33/common/crypto"
-
 	"github.com/33cn/chain33/client"
 	"github.com/33cn/chain33/common/version"
 	"github.com/33cn/chain33/pluginmgr"
@@ -1285,26 +1281,6 @@ func TestGetParaTxByHeight(t *testing.T) {
 	q := client.QueueProtocol{}
 	_, err := q.GetParaTxByHeight(nil)
 	assert.NotNil(t, err)
-}
-
-func TestQueueProtocol_GetCryptoList(t *testing.T) {
-	q := client.QueueProtocol{}
-	list := q.GetCryptoList()
-	for _, driver := range list.Cryptos {
-		id := int(driver.TypeID)
-		require.Equal(t, crypto.GetType(driver.Name), id)
-		require.Equal(t, crypto.GetName(id), driver.Name)
-	}
-}
-
-func TestQueueProtocol_GetAddressDrivers(t *testing.T) {
-	q := client.QueueProtocol{}
-	list := q.GetAddressDrivers()
-	for _, driver := range list.Drivers {
-		d, err := address.LoadDriver(driver.TypeID, -1)
-		require.Nil(t, err)
-		require.Equal(t, d.GetName(), driver.Name)
-	}
 }
 
 func TestQueueProtocol_SendDelayTx(t *testing.T) {
