+++ rpc/grpchandler_test.go
@@ -749,13 +749,11 @@ func TestGrpc_GetServerTime(t *testing.T) {
 }
 
 func TestGrpc_GetCryptoList(t *testing.T) {
-	qapi.On("GetCryptoList").Return(nil)
 	_, err := g.GetCryptoList(getOkCtx(), nil)
 	assert.NoError(t, err)
 }
 
 func TestGrpc_GetAddressDrivers(t *testing.T) {
-	qapi.On("GetAddressDrivers").Return(nil)
 	_, err := g.GetAddressDrivers(getOkCtx(), nil)
 	assert.NoError(t, err)
 }
